Graceful shutdown on SIGTERM for both gRPC and HTTP.
Note that graceful shutdown is not lame ducking. With graceful shutdown
existing requests are drained but the listening socket is dropped
immediately upon signal, so we do not accommodate clients that need time
to notice that we are going away.

However, in Kubernetes lame ducking may not be necessary as we are
probably removed from backend pools already before we get SIGTERM.
Still, might some clients be lagging behind and still try talking to
us, fail and not retry, and thus benefit if we lame duck? I don't
know, see https://github.com/hyperium/tonic/issues/1940