Major refactor of Assembly and Resource using lessons learned so far.
Resource will be replaced with a new, incompatible version, so rename it
to v0::Resource and prepare the Assembly for supporting multiple
different resource types. That's important since not all resources will
be able to migrate at the same time. Furthermore, I want to future-proof
Comprehensive by making sure adding a new resource type is possible
without disruption in case that is needed sometime after the API is
declared stable.

The new structure is based on a ResourceBase trait, which will form
the interface between the Assembly and various Resource types. The
ResourceBase trait subject to change but that's fine since it is sealed.

Another benefit of the refactor is that now a fully flexible topological
sort is used to calculate the Resource build order, which will surely
make things easier once trait-type dependencies (which is to say
requesting a dependency on any Resource present in the assembly that
implement some trait, instead of naming a concrete type to depend on)
are introduced.