Make ShutdownSignal safer and pass it to resources earlier.
Move ShutdownSignal and everything associated into its own module.

Constructing more than one ShutdownSignalParticipant for the same row
index could lead to reference counts becoming wrong, so make that
impossible by strictly controlling that object's creation using a consuming
iterator.

Pass ShutdownSignalParticipant objects for each resource earlier,
through new() instead of task(), since we have it available then and
this will make for a better v1 API.

Supply more than one ShutdownSignalParticipant for each resource so that
we can offer one to be owned by the resource and keep the other one for
cleanup in case it didn't. This, however, makes it possible to call
.propagate() more than once (which was not possible before, since the
method consumes self) which is bad for the refcounts as mentioned
before, so make sure that calls after the first are no-ops by reserving
the last reference cout for this. To make sure that method is only
available once the reference count has dropped enough, only offer it on
a struct that comes out of the Future that resolves when all other
references are gone.