Major incompatible changes to TLS and SPIFFE.
TlsConfig and TlsConfigProviders now exchange Box<dyn TlsConfigInstance>
objects instead of a simplistic snapshot of key, cert, and cacert. This
allows more complex identity selection and peer verification to happen
such as different roots for different SPIFFE domains. This also supports
having more than one TLS config provider at the same time.

The interface for TLS config providers moves from
comprehensive_traits::tls_config to comprehensive_tls::api. The idea
oof having the API in a separate crate was that if the API is fairly
low-dep then it allows implementors to import it cheaply. But it
really isn't low dep and most everything needs to depend on
comprehensive_tls anyway so it seemed a poor tradeoff.

From the perspective of a TLS client such as comprehensive_grpc there
are now unfortunately many notions of a URI at different protocol
layers. Roughly 4 of them wind up being exposed now:

1. The connection URI: provides the network address to connect to.
2. The HTTP-level URI: goes into HTTP Host / HTTP2 authority and
   also maybe SNI. This is considered the "main" URI (if you only
   configure one, make it this one, usually) and gives its value
   to #1 and #3 by default. But if #1 and #3 are both overridden,
   this one becomes pretty unimportant except maybe for proxies.
3. The expected server identity. If https: then it should match
   a DNS SAN. If spiffe: then it should match a URI SAN. Using a
   spiffe: URI only makes sense in this position and when one is
   used then at least one of #1 or #2 must surely be given with
   some other scheme like https:.
4. The preferred client identity. Useful if more than one is
   available and automatic selection using hints chooses wrong.

It remains to be seen whether this kind of 4-URI model for TLS
clients will be workable enough in practice.