A better definition of ResourceDependencies based on a trait for each dep.
This is a little more rigorous with types and also allows for a blanket
impl of ResourceDependencies on tuples.

Since there are different Resource traits (v0, v1, test) and I want to
keep that state as future-proofing, there is always some difficulty in
defining a common trait for all resources stemming from conflicting
blanket impls. With this change I abandon the earlier attempt to mitigate
that problem using a trait with a const generic. It was unwieldy and
never quite got all the way there. Instead, each Resource now has to
manually implement the common trait AnyResource. That's not so bad for
v1 resources which are all meant to be implemented with an attribute
macro (so the macro can impl it) but not so great for v0. Well, v0 is
on the way out. Anyway, maybe negative impls will help someday. The
new technique is heavily inspired by
https://www.greyblake.com/blog/alternative-blanket-implementations-for-single-rust-trait/

This change does not completely remove the old system because this
change is already big enough. Notably, there is a #[old_style]
attribute in the ResourceDependencies derive macro which allows the old
machinery to be selectively invoked.