Only Arc::clone exactly as many times as required.
It is a common pattern for one Resource to be the only one that depends
on another one, and in this case it might be handy for the first
Resource to have ownership of (or at least a mutable reference to) the
second. Arc::unwrap_or_clone is a good way to get that opportunistically
when possible but it doesn't work (in the sense that it will always
clone) if the Assembly is holding on to an unnecessary Arc reference
during Resource construction. So try to not do that.